+++ src/mpi/romio/adio/common/heap_sort.c
@@ -127,7 +127,7 @@ static void print_heap(heap_t * heap)
     printf("heap->size = %d\n", heap->size);
     printf("offsets:\n");
     for (i = 0; i < heap->size; i++) {
-        printf("%lld ", heap->nodes[i].offset);
+        printf("%lld ", (long long) heap->nodes[i].offset);
 
         if ((i + 1) == next_level_idx) {
             printf("\n");
